Issue #111: requested changes
